backend/requirements.txt
# WAR ROOM Backend Dependencies
openai
boto3
google-cloud-firestore
google-cloud-pubsub
fastapi
uvicorn[standard]
websockets
pydantic
pydantic-settings
python-dotenv
livekit-agents
livekit-plugins-aws[realtime]
livekit-plugins-google

# Testing
pytest
pytest-asyncio
